refactor add widgets for add sights scr to own directory
