+++ lib/ui/screen/add_sight_screen/add_sight_screen.dart
@@ -10,11 +10,11 @@ import 'package:places/data/model/sights/sight.dart';
 import 'package:places/data/model/enums/sight_type.dart';
 import 'package:places/mocks.dart';
 import 'package:places/styles/custom_icons.dart';
-import 'package:places/ui/screen/add_sight_screen/add_photo_pick_source_page.dart';
-import 'package:places/ui/screen/add_sight_screen/add_sight_photo_list.dart';
-import 'package:places/ui/screen/add_sight_screen/app_text_form_field.dart';
-import 'package:places/ui/screen/add_sight_screen/choose_category_screen.dart';
-import 'package:places/ui/screen/add_sight_screen/coordinate_text_field.dart';
+import 'package:places/ui/screen/add_sight_screen/screen_widgets/add_photo_pick_source_page.dart';
+import 'package:places/ui/screen/add_sight_screen/screen_widgets/add_sight_photo_list.dart';
+import 'package:places/ui/screen/add_sight_screen/screen_widgets/app_text_form_field.dart';
+import 'package:places/ui/screen/add_sight_screen/screen_widgets/choose_category_screen.dart';
+import 'package:places/ui/screen/add_sight_screen/screen_widgets/coordinate_text_field.dart';
 import 'package:places/ui/widgets/buttons/large_app_button.dart';
 import 'package:provider/provider.dart';
