+++ lib/ui/screen/add_sight_screen/screen_widgets/add_photo_button.dart
