+++ lib/ui/screen/add_sight_screen/screen_widgets/add_photo_pick_source_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:places/constants/app_strings.dart';
 import 'package:places/styles/custom_icons.dart';
-import 'package:places/ui/screen/add_sight_screen/pick_photo_source_row.dart';
+import 'package:places/ui/screen/add_sight_screen/screen_widgets/pick_photo_source_row.dart';
 import 'package:places/ui/widgets/buttons/large_app_button.dart';
 
 enum PickPhotoSourceType {
