+++ lib/ui/screen/add_sight_screen/screen_widgets/added_photo.dart
