+++ lib/ui/screen/add_sight_screen/screen_widgets/choose_category_screen.dart
