+++ lib/ui/screen/add_sight_screen/screen_widgets/pick_photo_source_row.dart
